Extract FFI call handling in executor into `FfiCall` data structure
